created files for RenterProfile subcomponents
